+++ c_client/subspace.cc
@@ -579,6 +579,7 @@ subspace_create_subscriber(SubspaceClient client, const char *channel_name,
                            SubspaceSubscriberOptions options) {
   subspace::SubscriberOptions subspace_options;
   subspace_options.SetReliable(options.reliable)
+      .SetLocal(options.local)
       .SetSubscriberQueueSize(options.subscriber_queue_size)
       .SetBridge(options.bridge)
       .SetForTunnel(options.for_tunnel)