+++ server/shadow_replicator.h
@@ -41,6 +41,7 @@ struct RecoveredPublisher {
 struct RecoveredSubscriber {
   int id = 0;
   bool is_reliable = false;
+  bool is_local = false;
   bool is_bridge = false;
   bool for_tunnel = false;
   int max_active_messages = 0;